moved shared functions to base class
